+++ boa/src/builtins/error/mod.rs
@@ -39,7 +39,7 @@ impl Error {
     pub(crate) const NAME: &'static str = "Error";
 
     /// The amount of arguments this function object takes.
-    pub(crate) const LENGTH: i32 = 1;
+    pub(crate) const LENGTH: usize = 1;
 
     /// Create a new error object.
     pub(crate) fn make_error(this: &mut Value, args: &[Value], _: &mut Interpreter) -> ResultValue {
@@ -94,6 +94,7 @@ impl Error {
     }
 
     /// Initialise the global object with the `Error` object.
+    #[inline]
     pub(crate) fn init(global: &Value) -> (&str, Value) {
         let _timer = BoaProfiler::global().start_event(Self::NAME, "init");
