+++ boa/src/builtins/error/range.rs
@@ -29,7 +29,7 @@ impl RangeError {
     pub(crate) const NAME: &'static str = "RangeError";
 
     /// The amount of arguments this function object takes.
-    pub(crate) const LENGTH: i32 = 1;
+    pub(crate) const LENGTH: usize = 1;
 
     /// Create a new error object.
     pub(crate) fn make_error(this: &mut Value, args: &[Value], _: &mut Interpreter) -> ResultValue {
@@ -84,6 +84,7 @@ impl RangeError {
     }
 
     /// Initialise the global object with the `RangeError` object.
+    #[inline]
     pub(crate) fn init(global: &Value) -> (&str, Value) {
         let _timer = BoaProfiler::global().start_event(Self::NAME, "init");
